+++ src/library/DIPS.Mobile.UI/Effects/Accessibility/Effects/TraitEffect.cs
@@ -1,0 +1,3 @@
+namespace DIPS.Mobile.UI.Effects.Accessibility.Effects;
+
+internal class TraitEffect : RoutingEffect;