+++ src/library/DIPS.Mobile.UI/Resources/LocalizedStrings/LocalizedStrings/DUILocalizedStrings.Designer.cs
@@ -464,5 +464,35 @@ namespace DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings {
                 return ResourceManager.GetString("NotSelected", resourceCulture);
             }
         }
+        
+        internal static string Accessability_SelectSortOption {
+            get {
+                return ResourceManager.GetString("Accessability_SelectSortOption", resourceCulture);
+            }
+        }
+        
+        internal static string Ascending {
+            get {
+                return ResourceManager.GetString("Ascending", resourceCulture);
+            }
+        }
+        
+        internal static string Descending {
+            get {
+                return ResourceManager.GetString("Descending", resourceCulture);
+            }
+        }
+        
+        internal static string Accessability_SelectedSort {
+            get {
+                return ResourceManager.GetString("Accessability_SelectedSort", resourceCulture);
+            }
+        }
+        
+        internal static string Accessability_TapToChangeSorting {
+            get {
+                return ResourceManager.GetString("Accessability_TapToChangeSorting", resourceCulture);
+            }
+        }
     }
 }